fixed view changes

+++ products/views.py
@@ -1,12 +1,13 @@
 from django.http import Http404,HttpResponse
 from django.shortcuts import render, redirect, get_object_or_404
 from .forms import AddCategoryForm
-from .models import Category
+from .models import Category, Product
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.contrib import messages
 from products.headings import AdminPortalHeadings
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.decorators import login_required
+from products.exceptions import CannotDeleteBrandException
 
 
 # Create your views here.
@@ -22,7 +23,7 @@ def add_category(request):
         category = AddCategoryForm(request.POST, request.FILES)
         if category.is_valid():
             category.save()
-            messages.success(request, AdminPortalHeadings.PRODUCT_ADDED)
+            messages.success(request, AdminPortalHeadings.CATEGORY_ADDED)
         return redirect('view-category')
 
     category = AddCategoryForm()
@@ -44,7 +45,7 @@ def update_category(request, pk):
         category = AddCategoryForm(request.POST, request.FILES, instance=category_instance)
         if category.is_valid():
             category.save()
-            messages.success(request, AdminPortalHeadings.PRODUCT_UPDATED)
+            messages.success(request, AdminPortalHeadings.CATEGORY_UPDATED)
         return redirect('view-category')
 
     category = AddCategoryForm(instance=category_instance)
@@ -85,9 +86,18 @@ def delete_category(request, pk):
 
     categorydata = get_object_or_404(Category, pk=pk)
     if request.method == "POST":
-
-        categorydata.delete()
-        messages.success(request, AdminPortalHeadings.PRODUCT_DELETED)
+        product = Product.objects.filter(category=pk)
+        try:
+            if not product:
+                categorydata.delete()
+                messages.success(request, AdminPortalHeadings.CATEGORY_DELETED)
+            else:
+                raise CannotDeleteBrandException
+        except CannotDeleteBrandException:
+            messages.info(request, AdminPortalHeadings.CATEGORY_NOT_DELETED)
         return redirect('view-category')
     else:
         return render(request, 'product/category/confirm_delete.html', {'category': categorydata})
+
+
+
